+++ openbank-admin-ui/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [0.119.0](https://github.com/JiRaska/open-bank-oss/compare/admin-ui-v0.118.0...admin-ui-v0.119.0) (2026-08-13)
+
+
+### Features
+
+* **campaign:** add delivery decision paths ([#4686](https://github.com/JiRaska/open-bank-oss/issues/4686)) ([73b2a0a](https://github.com/JiRaska/open-bank-oss/commit/73b2a0a82857b4676b3b613d9f88b73b94416c11))
+
+## [0.118.0](https://github.com/JiRaska/open-bank-oss/compare/admin-ui-v0.117.0...admin-ui-v0.118.0) (2026-08-13)
+
+
+### Features
+
+* **campaign:** let makers revise drafts ([#4667](https://github.com/JiRaska/open-bank-oss/issues/4667)) ([60b86bb](https://github.com/JiRaska/open-bank-oss/commit/60b86bb48e7c480bd05b79a5c8a2c544f9e2b1f4))
+
+## [0.117.0](https://github.com/JiRaska/open-bank-oss/compare/admin-ui-v0.116.0...admin-ui-v0.117.0) (2026-08-13)
+
+
+### Features
+
+* **admin-ui:** add governed audience library ([#4653](https://github.com/JiRaska/open-bank-oss/issues/4653)) ([f2dc472](https://github.com/JiRaska/open-bank-oss/commit/f2dc472990b45978e4f0486099ac4589404cb08e)), closes [#4652](https://github.com/JiRaska/open-bank-oss/issues/4652)
+
+## [0.116.0](https://github.com/JiRaska/open-bank-oss/compare/admin-ui-v0.115.0...admin-ui-v0.116.0) (2026-08-13)
+
+
+### Features
+
+* **admin-ui:** add app-first journey recipes ([#4643](https://github.com/JiRaska/open-bank-oss/issues/4643)) ([fa06243](https://github.com/JiRaska/open-bank-oss/commit/fa06243b5673dd0b06e3efe175a481bc66c495c9))
+
 ## [0.115.0](https://github.com/JiRaska/open-bank-oss/compare/admin-ui-v0.114.1...admin-ui-v0.115.0) (2026-08-13)